// client/src/app/services/pipette/pipette-properties.ts
export class PipetteProperties {
  color: string;
  centerX: number;
  centerY: number;
  ajustedPointsX: number[];
  ajustedPointsY: number[];
  toolbarWidth: number;
  amountScrolledX = 0;
  amountScrolledY = 0;
  primaryColor: string;
  secondaryColor: string;
  positionX: number;
  positionY: number;
}